Capitalisation correction and modified two holidays
As far as I can tell, the capitalisation is now correct.  The only one not with
capitals is the "Premier lundi d'août" in fr-CA and as far as I could find, that
is the correct case.

I have removed Manitoba from family day, since I have found out it is not
actually an official holiday in Manitoba.

I have added Saskatchewan on the first Monday of August list, since although
it is not the "Civic Holiday" of the other provinces, I have discovered it is
Saskatchewan Day.